[WIP] Remove info sidebar detail
