+++ app/src/modules/insights/routes/dashboard.vue
@@ -283,10 +283,6 @@ const refreshInterval = computed({
 		</template>
 
 		<template #sidebar>
-			<sidebar-detail icon="info" :title="t('information')" close>
-				<div v-md="t('page_help_insights_dashboard')" class="page-description" />
-			</sidebar-detail>
-
 			<comments-sidebar-detail :key="primaryKey" collection="directus_dashboards" :primary-key="primaryKey" />
 
 			<refresh-sidebar-detail v-model="refreshInterval" @refresh="insightsStore.refresh(primaryKey)" />
