--- app/src/modules/settings/routes/extensions/components/extensions-info-sidebar-detail.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-import { useI18n } from 'vue-i18n';
-
-const { t } = useI18n();
-</script>
-
-<template>
-	<sidebar-detail icon="info" :title="t('information')" close>
-		<div v-md="t('page_help_settings_extensions')" class="page-description" />
-	</sidebar-detail>
-</template>
