+++ app/src/modules/settings/routes/extensions/extensions.vue
@@ -42,10 +42,6 @@ const extensionsByType = computed(() => {
 			<settings-navigation />
 		</template>
 
-		<template #sidebar>
-			<extensions-info-sidebar-detail />
-		</template>
-
 		<div v-if="extensions.length > 0 || loading === false" class="page-container">
 			<template v-if="extensions.length > 0">
 				<div v-for="(list, type) in extensionsByType" :key="`${type}-list`" class="extension-group">
