+++ app/src/modules/settings/routes/flows/overview.vue
@@ -149,12 +149,6 @@ function onFlowDrawerCompletion(id: string) {
 			</v-button>
 		</template>
 
-		<template #sidebar>
-			<sidebar-detail icon="info" :title="t('information')" close>
-				<div v-md="t('page_help_settings_flows_collection')" class="page-description" />
-			</sidebar-detail>
-		</template>
-
 		<v-info v-if="flows.length === 0" icon="bolt" :title="t('no_flows')" center>
 			{{ t('no_flows_copy') }}
