--- app/src/modules/settings/routes/marketplace/routes/account/components/account-info-sidebar-detail.vue
@@ -1,11 +1,0 @@
-<script setup lang="ts">
-import { useI18n } from 'vue-i18n';
-
-const { t } = useI18n();
-</script>
-
-<template>
-	<sidebar-detail icon="info" :title="t('information')" close>
-		<div v-md="t('page_help_marketplace_account')" />
-	</sidebar-detail>
-</template>
