+++ app/src/modules/settings/routes/roles/role-info-sidebar-detail.vue
@@ -13,7 +13,7 @@ const { isCopySupported, copyToClipboard } = useClipboard();
 </script>
 
 <template>
-	<sidebar-detail icon="info" :title="t('information')" close>
+	<sidebar-detail id="role" icon="info" :title="t('information')" close>
 		<template v-if="role">
 			<dl>
 				<div class="description-list">
