+++ app/src/modules/settings/routes/webhooks/collection.vue
@@ -128,9 +128,6 @@ function clearFilters() {
 			</component>
 
 			<template #sidebar>
-				<sidebar-detail icon="info" :title="t('information')" close>
-					<div v-md="t('page_help_settings_webhooks_collection')" class="page-description" />
-				</sidebar-detail>
 				<layout-sidebar-detail v-model="layout">
 					<component :is="`layout-options-${layout}`" v-bind="layoutState" />
 				</layout-sidebar-detail>
