+++ app/src/views/private/components/basic-import-sidebar-detail.vue
@@ -103,7 +103,7 @@ function useUpload() {
 </script>
 
 <template>
-	<sidebar-detail icon="publish" :title="t('label_import')">
+	<sidebar-detail id="import" icon="publish" :title="t('label_import')">
 		<div class="fields">
 			<template v-if="createAllowed">
 				<div class="field full">
