+++ app/src/views/private/components/flow-sidebar-detail.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import { useI18n } from 'vue-i18n';
-import { FlowRaw } from '@directus/types';
 import { injectRunManualFlow } from '@/composables/use-flows';
+import { FlowRaw } from '@directus/types';
+import { useI18n } from 'vue-i18n';
 
 const { t } = useI18n();
 
@@ -17,7 +17,7 @@ const { runManualFlow } = injectRunManualFlow();
 </script>
 
 <template>
-	<sidebar-detail v-if="manualFlows.length > 0" icon="bolt" :title="t('flows')">
+	<sidebar-detail v-if="manualFlows.length > 0" id="flows" icon="bolt" :title="t('flows')">
 		<div class="fields">
 			<div v-for="manualFlow in manualFlows" :key="manualFlow.id" class="field full">
 				<v-button
