+++ app/src/views/private/components/refresh-sidebar-detail.vue
@@ -66,7 +66,7 @@ const items = computed(() => {
 </script>
 
 <template>
-	<sidebar-detail :icon="active ? 'sync' : 'sync_disabled'" :title="t('auto_refresh')" :badge="active">
+	<sidebar-detail id="refresh" :icon="active ? 'sync' : 'sync_disabled'" :title="t('auto_refresh')" :badge="active">
 		<div class="fields">
 			<div class="field full">
 				<p class="type-label">{{ t('refresh_interval') }}</p>
