+++ app/src/views/private/components/revisions-sidebar-detail.vue
@@ -78,6 +78,7 @@ defineExpose({
 
 <template>
 	<sidebar-detail
+		id="revisions"
 		:title
 		icon="change_history"
 		:badge="!loadingCount && revisionsCount > 0 ? abbreviateNumber(revisionsCount) : null"
