+++ app/src/views/private/private-view/components/private-view-sidebar.vue
@@ -1,20 +1,14 @@
 <script lang="ts" setup>
-import SidebarDetailGroup from './private-view-sidebar-detail-group.vue';
 import SkipMenu from '../../components/skip-menu.vue';
+import { AccordionRoot } from 'reka-ui';
 </script>
 
 <template>
 	<aside id="sidebar" ref="sidebarEl" role="contentinfo" class="alt-colors" aria-label="Module Sidebar">
 		<SkipMenu section="sidebar" />
-		<div class="flex-container">
-			<SidebarDetailGroup>
-				<slot name="sidebar" />
-			</SidebarDetailGroup>
-
-			<div class="spacer" />
-
-			<!-- TODO <notifications-preview v-model="notificationsPreviewActive" :sidebar-open="sidebarOpen" /> -->
-		</div>
+		<AccordionRoot type="single" collapsible>
+			<slot name="sidebar" />
+		</AccordionRoot>
 	</aside>
 </template>
 
@@ -26,25 +20,9 @@ import SkipMenu from '../../components/skip-menu.vue';
 	background-color: var(--theme--sidebar--background);
 	font-family: var(--theme--sidebar--font-family);
 	border-inline-start: var(--theme--sidebar--border-width) solid var(--theme--sidebar--border-color);
-	min-inline-size: 220px;
+	min-inline-size: 250px;
 
 	/* Explicitly render the border outside of the width of the bar itself */
 	box-sizing: content-box;
-
-	.spacer {
-		flex-grow: 1;
-	}
-
-	/* TODO */
-	&.has-shadow {
-		box-shadow: var(--sidebar-shadow);
-	}
-
-	.flex-container {
-		display: flex;
-		flex-direction: column;
-		inline-size: 100%;
-		block-size: 100%;
-	}
 }
 </style>
